requirements.txt
Django~=2.2
pytz
djangorestframework
django-rest-passwordreset
requests
pyyaml
drf-yasg